+++ cedar-policy-symcc/CHANGELOG.md
@@ -10,7 +10,10 @@ Cedar Language Version: TBD
 
 ### Fixed
 
-- Model decoder now accepts hexadecimal bitvectors in solver output, improving compatibility with Z3 and other solvers that include literals like `#xff` in their `(get-model)` responses.
+- Fix errors decoding models from Z3. The model decoder now accepts hexadecimal bitvectors
+  in solver output (e.g., `#xff`). It also handles an option literal `none`
+  without an explicit type annotation, as long as it can infer the type. Finally, it accepts
+  models for UUFs with equality operands in either order.
 
 ## [0.5.2] - 2026-06-09
 Cedar Language Version: 4.5